XILLOADER:Fix MISRA-C 10.7 Violation
MISRA-C Rule 10.7:If a composite expression is used as one operand of an operator in which the usual arithmetic conversions are performed then the other operand shall not have wider essential type

Solution:Typecasted accordingly


Acked-by: Preethi Kolli <preethi.kolli@amd.com>

+++ lib/sw_services/xilloader/src/versal_net/server/xloader_plat.c
@@ -1911,7 +1911,7 @@ int XLoader_StoreAppVersion(u32 OptionalDataLen, u32 OptionalDataId)
 {
 	int Status = XST_FAILURE;
 	u64 OptionalDataStartAddr = XILPDI_PMCRAM_IHT_DATA_ADDR;
-	u64 OptionalDataEndAddr = OptionalDataStartAddr + (OptionalDataLen << XILPDI_WORD_LEN_SHIFT);
+	u64 OptionalDataEndAddr = OptionalDataStartAddr + ((u64)OptionalDataLen << XILPDI_WORD_LEN_SHIFT);
 	u64 OptDataAddr;
 	u64 OptDataHdr;
 	u32 OptDataLen;
